Refactor: 모듈화 진행하여 코드 구조 개선
- FastAPI 애플리케이션을 여러 파일로 분리하여 가독성 향상
- 요청 핸들러, 유틸리티 함수, Pydantic 모델, 문서 로더 등으로 파일 구성
- 각 기능 별로 모듈화하여 유지보수 용이성 증가
