+++ models.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class UserInput(BaseModel):
+    user_input: str
